greader_compact: Switch to low-contrast color scheme
Make unread items have a light-grey font, so they're a lot lower
contrast. When reading a page-full of short articles (e.g.
Slickdeals), having all the bold-black text at the bottom is pretty
distracting. This way, the expanded article has the most visual
weight and is the only thing that is black-text.